:recycle: refactor: all hooks refactored
useCars splited on more hooks and updates to useReducer implementation
