+++ web/src/hooks/useDeleteCars.js
@@ -1,6 +1,6 @@
 import api from "../services/api";
 
-export default async function useDeleteCars() {
+export default function useDeleteCars() {
   async function handleDeleteCar(carId, callback) {
     await api("DELETE", `/cars/${carId}`).then((r) => callback(r));
   }
