+++ web/src/hooks/usePostCars.js
@@ -1,6 +1,6 @@
 import api from "../services/api";
 
-export default async function usePostCars(carData) {
+export default function usePostCars(carData) {
   async function handlePostCars(callback) {
     await api("POST", "/cars", carData).then((r) => callback(r));
   }
