+++ web/src/hooks/usePutCars.js
@@ -1,6 +1,6 @@
 import api from "../services/api";
 
-export default async function usePutCars(carData) {
+export default function usePutCars(carData) {
   async function handlePutUpdateCar(carId, callback) {
     await api("PUT", `/cars/${carId}`, carData).then((r) => callback(r));
   }
